c10

+++ C07/ex01/ft_range.c
@@ -6,7 +6,7 @@
 /*   By: jrinta- <jrinta-@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/07/05 05:56:47 by jrinta-           #+#    #+#             */
-/*   Updated: 2024/07/09 18:30:15 by jrinta-          ###   ########.fr       */
+/*   Updated: 2024/07/17 17:29:12 by jrinta-          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ int	*ft_range(int min, int max)
 
 	if (min >= max)
 		return (0);
-	range = (long)max - (long)min;
+	range = (unsigned)max - (unsigned)min;
 	if (range < 0)
 		range = -range;
 	str = malloc(range * 4);
